moved prenormalizer to console controller
